+++ plugins/index.ts
@@ -4,15 +4,17 @@ import vueJsx from '@vitejs/plugin-vue-jsx'
 import AutoImport from 'unplugin-auto-import/vite'
 import GenerateConfig from 'unplugin-config/vite'
 import Components from 'unplugin-vue-components/vite'
+import VitePluginPreloadAll from '@mistjs/vite-plugin-preload'
 import Unocss from 'unocss/vite'
 import AntdvResolver from 'antdv-component-resolver'
-import { APP_NAME, GLOB_CONFIG_FILE_NAME, OUTPUT_DIR } from './constants'
+import { GLOB_CONFIG_FILE_NAME, OUTPUT_DIR } from './constants'
 import { viteBuildInfo } from './vite-build-info'
 
 export function createVitePlugins(env: Record<string, string>) {
   const vitePluginList: (PluginOption | PluginOption[])[] = [
     vue(),
     vueJsx(),
+    VitePluginPreloadAll(),
     AutoImport({
       imports: [
         'vue',
@@ -31,7 +33,7 @@ export function createVitePlugins(env: Record<string, string>) {
     }),
     // https://github.com/kirklin/unplugin-config
     GenerateConfig({
-      appName: APP_NAME,
+      appName: env.VITE_GLOB_APP_TITLE,
       configFile: {
         generate: true,
         fileName: GLOB_CONFIG_FILE_NAME,