+++ src/layouts/index.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 import { pick } from '@v-c/utils'
-import { defineOptions } from 'vue'
 import BasicLayout from './basic-layout/index.vue'
 import SettingDrawer from './components/setting-drawer/index.vue'
 import MultiTab from './multi-tab/index.vue'
@@ -32,6 +31,7 @@ const layoutProps = computed(() =>
     'layout',
     'footer',
     'contentWidth',
+    'compactAlgorithm',
   ]),
 )
 </script>
@@ -84,6 +84,7 @@ const layoutProps = computed(() =>
     :theme="layoutSetting.theme"
     :color-primary="layoutSetting.colorPrimary"
     :color-weak="layoutSetting.colorWeak"
+    :color-gray="layoutSetting.colorGray"
     :multi-tab="layoutSetting.multiTab"
     :multi-tab-fixed="layoutSetting.multiTabFixed"
     :animation-name-list="animationNameList"