+++ src/stores/app.ts
@@ -22,6 +22,7 @@ export interface LayoutSetting {
   menu?: boolean
   menuHeader?: boolean
   colorWeak?: boolean
+  colorGray?: boolean
   multiTab?: boolean
   multiTabFixed?: boolean
   headerHeight?: number
@@ -29,10 +30,12 @@ export interface LayoutSetting {
   keepAlive?: boolean
   accordionMode?: boolean
   leftCollapsed?: boolean
+  compactAlgorithm?: boolean
   animationName?: AnimationNameValueType
 }
 
 export const useAppStore = defineStore('app', () => {
+  const { darkAlgorithm, compactAlgorithm, defaultAlgorithm } = antdTheme
   const layoutSetting = reactive<LayoutSetting>(defaultSetting)
   const themeConfig = reactive<ThemeConfig>({
     algorithm: antdTheme.defaultAlgorithm,
@@ -46,6 +49,12 @@ export const useAppStore = defineStore('app', () => {
   const toggleLocale = (locale: string) => {
     lsLocaleState.value = locale
   }
+  const toggleCompact = (isCompact = true) => {
+    layoutSetting.compactAlgorithm = isCompact
+    const algorithm = layoutSetting.theme === 'dark' ? [darkAlgorithm] : [defaultAlgorithm]
+    isCompact && algorithm.push(compactAlgorithm)
+    themeConfig.algorithm = algorithm
+  }
   const toggleTheme = (theme: ThemeType) => {
     if (layoutSetting.theme === theme)
       return
@@ -55,9 +64,6 @@ export const useAppStore = defineStore('app', () => {
         themeConfig.token.colorBgContainer = '#fff'
       if (themeConfig.components?.Menu)
         delete themeConfig.components.Menu
-
-      themeConfig.algorithm = antdTheme.defaultAlgorithm
-
       toggleDark(false)
     }
     else if (theme === 'dark') {
@@ -74,8 +80,8 @@ export const useAppStore = defineStore('app', () => {
           } as any,
         }
       }
-      themeConfig.algorithm = antdTheme.darkAlgorithm
     }
+    toggleCompact(layoutSetting.compactAlgorithm)
   }
 
   const toggleDrawerVisible = (visible: boolean) => {
@@ -108,6 +114,34 @@ export const useAppStore = defineStore('app', () => {
     layoutSetting.collapsed = collapsed
   }
 
+  function toggleGray(isGray = true) {
+    layoutSetting.colorGray = isGray
+    const dom = document.querySelector('body')
+    if (dom) {
+      if (isGray) {
+        toggleWeak(false)
+        dom.style.filter = 'grayscale(100%)'
+      }
+      else { dom.style.filter = '' }
+    }
+  }
+  if (layoutSetting.colorGray)
+    toggleGray(true)
+
+  function toggleWeak(isWeak = true) {
+    layoutSetting.colorWeak = isWeak
+    const dom = document.querySelector('body')
+    if (dom) {
+      if (isWeak) {
+        toggleGray(false)
+        dom.style.filter = 'invert(80%)'
+      }
+      else { dom.style.filter = '' }
+    }
+  }
+  if (layoutSetting.colorWeak)
+    toggleWeak(true)
+
   const toggleLayout = (layout: LayoutType) => {
     if (layoutSetting.theme === 'inverted' && layout === 'mix')
       layoutSetting.theme = 'light'
@@ -130,6 +164,12 @@ export const useAppStore = defineStore('app', () => {
       toggleColorPrimary(value)
     else if (key === 'layout')
       toggleLayout(value as LayoutType)
+    else if (key === 'compactAlgorithm')
+      toggleCompact(value)
+    else if (key === 'colorWeak')
+      toggleWeak(value)
+    else if (key === 'colorGray')
+      toggleGray(value)
     else if (key in layoutSetting)
       (layoutSetting as any)[key] = value
   }
@@ -144,5 +184,7 @@ export const useAppStore = defineStore('app', () => {
     toggleDrawerVisible,
     changeSettingLayout,
     toggleColorPrimary,
+    toggleGray,
+    toggleWeak,
   }
 })