+++ src/stores/user.ts
@@ -3,7 +3,7 @@ import { getRouteMenusApi } from '~@/api/common/menu'
 import type { UserInfo } from '~@/api/common/user'
 import { getUserInfoApi } from '~@/api/common/user'
 import type { MenuData } from '~@/layouts/basic-layout/typing'
-import { rootRoute } from '~@/router/dynamic-routes'
+import { rootRoute } from '~@/router/constant'
 import { generateFlatRoutes, generateRoutes, generateTreeRoutes } from '~@/router/generate-route'
 import { DYNAMIC_LOAD_WAY, DynamicLoadEnum } from '~@/utils/constant'