+++ unocss.config.ts
@@ -4,7 +4,6 @@ import {
   presetIcons,
   presetTypography,
   presetUno,
-  presetWebFonts,
   transformerDirectives,
   transformerVariantGroup,
 } from 'unocss'
@@ -13,7 +12,10 @@ import presetEase from 'unocss-preset-ease'
 import antdUnoTheme from './themes/antd-uno-theme.json'
 
 export default defineConfig({
-  safelist: [],
+  safelist: [
+    'py-16px',
+    'pb-16px',
+  ],
   theme: {
     ...antdUnoTheme,
   },
@@ -26,13 +28,6 @@ export default defineConfig({
     presetIcons({
       scale: 1.2,
       warn: true,
-    }),
-    presetWebFonts({
-      fonts: {
-        sans: 'DM Sans',
-        serif: 'DM Serif Display',
-        mono: 'DM Mono',
-      },
     }),
   ],
   shortcuts: [