+++ CHANGELOG.md
@@ -1,6 +1,179 @@
 # Changelog
 
 
+## v1.1.0
+
+[compare changes](https://github.com/antdv-pro/antdv-pro/compare/v1.0.2...v1.1.0)
+
+### 🚀 Enhancements
+
+- Add table-query hook for crud table ([ab64028](https://github.com/antdv-pro/antdv-pro/commit/ab64028))
+
+### 🩹 Fixes
+
+- Eslint-config in vscode ([f8018ac](https://github.com/antdv-pro/antdv-pro/commit/f8018ac))
+- 路由切换可能导致page-container的头部面包屑等区域被意外隐藏 ([03d3ea8](https://github.com/antdv-pro/antdv-pro/commit/03d3ea8))
+- Nested not change ([afa71cd](https://github.com/antdv-pro/antdv-pro/commit/afa71cd))
+
+### 💅 Refactors
+
+- Page container ([0c0cd0d](https://github.com/antdv-pro/antdv-pro/commit/0c0cd0d))
+
+### 🏡 Chore
+
+- Remove icon ([79cb981](https://github.com/antdv-pro/antdv-pro/commit/79cb981))
+- Update deps ([4535f65](https://github.com/antdv-pro/antdv-pro/commit/4535f65))
+
+### ❤️ Contributors
+
+- Aibayanyu20 <aibayanyu@qq.com>
+- Sun1090 <1090964491@qq.com>
+- Wangqifan <wangqifan@wxdata.cn>
+
+## v1.0.2
+
+[compare changes](https://github.com/antdv-pro/antdv-pro/compare/v1.0.1...v1.0.2)
+
+### 🩹 Fixes
+
+- Typo ([ba8f474](https://github.com/antdv-pro/antdv-pro/commit/ba8f474))
+
+### 🏡 Chore
+
+- Update deps ([0d41da9](https://github.com/antdv-pro/antdv-pro/commit/0d41da9))
+- Use antfu eslint config replace mist config ([c11224e](https://github.com/antdv-pro/antdv-pro/commit/c11224e))
+- Update deps ([a12b56f](https://github.com/antdv-pro/antdv-pro/commit/a12b56f))
+
+### ❤️ Contributors
+
+- Aibayanyu20 <aibayanyu@qq.com>
+
+## v1.0.1
+
+[compare changes](https://github.com/antdv-pro/antdv-pro/compare/v1.0.0...v1.0.1)
+
+### 🩹 Fixes
+
+- Fix circular imports ([7ab99da](https://github.com/antdv-pro/antdv-pro/commit/7ab99da))
+- Fix page container dynamic update in advance ([13d9486](https://github.com/antdv-pro/antdv-pro/commit/13d9486))
+
+### 🏡 Chore
+
+- **release:** V1.0.0 ([e7ca880](https://github.com/antdv-pro/antdv-pro/commit/e7ca880))
+
+### ❤️ Contributors
+
+- Aibayanyu20 <aibayanyu@qq.com>
+
+## v1.0.0
+
+[compare changes](https://github.com/antdv-pro/antdv-pro/compare/v1.0.0-beta.3...v1.0.0)
+
+### 🚀 Enhancements
+
+- Add preload ([6c0e6c9](https://github.com/antdv-pro/antdv-pro/commit/6c0e6c9))
+- Support vite 5 ([e84c7b7](https://github.com/antdv-pro/antdv-pro/commit/e84c7b7))
+- Complete compact-algorithm ([99f2074](https://github.com/antdv-pro/antdv-pro/commit/99f2074))
+- Query table add toolbar ([7fe8f90](https://github.com/antdv-pro/antdv-pro/commit/7fe8f90))
+
+### 🩹 Fixes
+
+- Change ([8bdfd94](https://github.com/antdv-pro/antdv-pro/commit/8bdfd94))
+- Preload error ([c0de1fc](https://github.com/antdv-pro/antdv-pro/commit/c0de1fc))
+- Fix top mode page-container not expand close #127 ([#127](https://github.com/antdv-pro/antdv-pro/issues/127))
+- Eslint error ([39dbe6f](https://github.com/antdv-pro/antdv-pro/commit/39dbe6f))
+- Eslint not support ([d249d98](https://github.com/antdv-pro/antdv-pro/commit/d249d98))
+- Note ([92beb3e](https://github.com/antdv-pro/antdv-pro/commit/92beb3e))
+- X-axis scroll bar appears in top layout mode ([bf2a206](https://github.com/antdv-pro/antdv-pro/commit/bf2a206))
+- Footertoolbar not effect layout ([e025b9c](https://github.com/antdv-pro/antdv-pro/commit/e025b9c))
+
+### 🏡 Chore
+
+- Delete eslintignore replace eslint.config ([baa7747](https://github.com/antdv-pro/antdv-pro/commit/baa7747))
+- Update deps ([6fc6496](https://github.com/antdv-pro/antdv-pro/commit/6fc6496))
+- Change tsconfig ([734e287](https://github.com/antdv-pro/antdv-pro/commit/734e287))
+- Update scripts ([f372ac9](https://github.com/antdv-pro/antdv-pro/commit/f372ac9))
+- Change script ([1e80dd4](https://github.com/antdv-pro/antdv-pro/commit/1e80dd4))
+- Update deps ([0de46db](https://github.com/antdv-pro/antdv-pro/commit/0de46db))
+- Support ssl config ([ca9d37b](https://github.com/antdv-pro/antdv-pro/commit/ca9d37b))
+- Update deps ([b7c557c](https://github.com/antdv-pro/antdv-pro/commit/b7c557c))
+- Update deps ([ff86513](https://github.com/antdv-pro/antdv-pro/commit/ff86513))
+- Update deps ([abaee14](https://github.com/antdv-pro/antdv-pro/commit/abaee14))
+- Add funding ([03ef885](https://github.com/antdv-pro/antdv-pro/commit/03ef885))
+- Change ([ceb08eb](https://github.com/antdv-pro/antdv-pro/commit/ceb08eb))
+- Update vue@3.4.x ([a9e9ac9](https://github.com/antdv-pro/antdv-pro/commit/a9e9ac9))
+
+### 🤖 CI
+
+- Change node version ([540224b](https://github.com/antdv-pro/antdv-pro/commit/540224b))
+
+### ❤️ Contributors
+
+- Aibayanyu20 <aibayanyu@qq.com>
+- ZhouWei <1244620067@qq.com>
+- Undefined ([@undefined-moe](http://github.com/undefined-moe))
+- LinRenJie <xoxosos666@gmail.com>
+- Aibayanyu <aibayanyu@qq.com>
+
+## v1.0.0-beta.3
+
+[compare changes](https://github.com/antdv-pro/antdv-pro/compare/v1.0.0-beta.2...v1.0.0-beta.3)
+
+### 🩹 Fixes
+
+- Locale error ([df1293d](https://github.com/antdv-pro/antdv-pro/commit/df1293d))
+
+### 🏡 Chore
+
+- Update readme ([27ab14e](https://github.com/antdv-pro/antdv-pro/commit/27ab14e))
+- Update deps ([9946ea4](https://github.com/antdv-pro/antdv-pro/commit/9946ea4))
+- Update deps ([aae9b3a](https://github.com/antdv-pro/antdv-pro/commit/aae9b3a))
+- Unified build output directory #118 ([#118](https://github.com/antdv-pro/antdv-pro/issues/118))
+- Update deps ([75937d6](https://github.com/antdv-pro/antdv-pro/commit/75937d6))
+
+### ❤️ Contributors
+
+- Aibayanyu <aibayanyu@qq.com>
+- Aibayanyu20 <aibayanyu@qq.com>
+- Kirk Lin ([@kirklin](http://github.com/kirklin))
+
+## v1.0.0-beta.2
+
+[compare changes](https://github.com/antdv-pro/antdv-pro/compare/v1.0.0-beta.1...v1.0.0-beta.2)
+
+### 🚀 Enhancements
+
+- Add vitest ([ec66360](https://github.com/antdv-pro/antdv-pro/commit/ec66360))
+- Add icp ([ced7249](https://github.com/antdv-pro/antdv-pro/commit/ced7249))
+
+### 🩹 Fixes
+
+- Fix the page not scrolling to the top when switching ([d6a2406](https://github.com/antdv-pro/antdv-pro/commit/d6a2406))
+- Fix locale auto load error ([e441349](https://github.com/antdv-pro/antdv-pro/commit/e441349))
+- Using vite's outDir to unify staticPath ([d5f5448](https://github.com/antdv-pro/antdv-pro/commit/d5f5448))
+- Fix typo ([e68bb8c](https://github.com/antdv-pro/antdv-pro/commit/e68bb8c))
+- Fix i18n hot accept error ([e1da2a8](https://github.com/antdv-pro/antdv-pro/commit/e1da2a8))
+- Card layout misalignment ([1944820](https://github.com/antdv-pro/antdv-pro/commit/1944820))
+- Remove tabs default margin ([8de0518](https://github.com/antdv-pro/antdv-pro/commit/8de0518))
+- Not has multitab prcessing pagecontainer spacing ([a604f9d](https://github.com/antdv-pro/antdv-pro/commit/a604f9d))
+- Icp open blank page ([f0739e2](https://github.com/antdv-pro/antdv-pro/commit/f0739e2))
+- Page show not right close #122 ([#122](https://github.com/antdv-pro/antdv-pro/issues/122))
+
+### 🏡 Chore
+
+- Change version ([742270e](https://github.com/antdv-pro/antdv-pro/commit/742270e))
+- Update deps ([d291e93](https://github.com/antdv-pro/antdv-pro/commit/d291e93))
+- Update deps ([00b693e](https://github.com/antdv-pro/antdv-pro/commit/00b693e))
+
+### ❤️ Contributors
+
+- Aibayanyu <aibayanyu@qq.com>
+- Jiabochao ([@jiabochao](http://github.com/jiabochao))
+- Undefined ([@undefined-moe](http://github.com/undefined-moe))
+- 萍萍 ([@jiabochao](http://github.com/jiabochao))
+- Aibayanyu20 <aibayanyu@qq.com>
+- Unknown ([@aibayanyu20](http://github.com/aibayanyu20))
+
 ## v1.0.0-beta.1
 
 [compare changes](https://github.com/antdv-pro/antdv-pro/compare/v0.2.0...v1.0.0-beta.1)